[REFACTOR] refactor domain.auth
