+++ src/main/java/com/sullung2yo/seatcatcher/common/jwt/provider/JwtTokenProviderImpl.java
@@ -3,10 +3,10 @@ package com.sullung2yo.seatcatcher.common.jwt.provider;
 import com.sullung2yo.seatcatcher.common.exception.ErrorCode;
 import com.sullung2yo.seatcatcher.common.exception.TokenException;
 import com.sullung2yo.seatcatcher.common.domain.TokenType;
-import com.sullung2yo.seatcatcher.user.domain.RefreshToken;
+import com.sullung2yo.seatcatcher.domain.auth.entity.RefreshToken;
 import com.sullung2yo.seatcatcher.user.domain.User;
 import com.sullung2yo.seatcatcher.user.domain.UserRole;
-import com.sullung2yo.seatcatcher.user.repository.RefreshTokenRepository;
+import com.sullung2yo.seatcatcher.domain.auth.repository.RefreshTokenRepository;
 import com.sullung2yo.seatcatcher.user.repository.UserRepository;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
