+++ src/main/java/com/sullung2yo/seatcatcher/domain/auth/controller/AuthController.java
@@ -1,11 +1,11 @@
-package com.sullung2yo.seatcatcher.user.controller;
+package com.sullung2yo.seatcatcher.domain.auth.controller;
 
 
-import com.sullung2yo.seatcatcher.user.domain.Provider;
-import com.sullung2yo.seatcatcher.user.dto.request.AppleAuthRequest;
-import com.sullung2yo.seatcatcher.user.dto.request.KakaoAuthRequest;
-import com.sullung2yo.seatcatcher.user.dto.response.TokenResponse;
-import com.sullung2yo.seatcatcher.user.service.AuthServiceImpl;
+import com.sullung2yo.seatcatcher.domain.auth.enums.Provider;
+import com.sullung2yo.seatcatcher.domain.auth.dto.request.AppleAuthRequest;
+import com.sullung2yo.seatcatcher.domain.auth.dto.request.KakaoAuthRequest;
+import com.sullung2yo.seatcatcher.domain.auth.dto.response.TokenResponse;
+import com.sullung2yo.seatcatcher.domain.auth.service.AuthServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
