+++ src/main/java/com/sullung2yo/seatcatcher/domain/auth/dto/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.sullung2yo.seatcatcher.user.dto.response;
+package com.sullung2yo.seatcatcher.domain.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
