+++ src/main/java/com/sullung2yo/seatcatcher/domain/auth/entity/RefreshToken.java
@@ -1,7 +1,8 @@
-package com.sullung2yo.seatcatcher.user.domain;
+package com.sullung2yo.seatcatcher.domain.auth.entity;
 
 
 import com.sullung2yo.seatcatcher.common.domain.BaseEntity;
+import com.sullung2yo.seatcatcher.user.domain.User;
 import jakarta.persistence.*;
 import lombok.*;
