+++ src/main/java/com/sullung2yo/seatcatcher/domain/auth/enums/Provider.java
@@ -1,0 +1,5 @@
+package com.sullung2yo.seatcatcher.domain.auth.enums;
+
+public enum Provider {
+    LOCAL, KAKAO, APPLE
+}
