--- src/main/java/com/sullung2yo/seatcatcher/user/domain/Provider.java
@@ -1,5 +1,0 @@
-package com.sullung2yo.seatcatcher.user.domain;
-
-public enum Provider {
-    LOCAL, KAKAO, APPLE
-}
