+++ src/test/java/com/sullung2yo/seatcatcher/auth/AppleAuthTest.java
@@ -3,12 +3,12 @@ package com.sullung2yo.seatcatcher.auth;
 import com.sullung2yo.seatcatcher.common.exception.TokenException;
 import com.sullung2yo.seatcatcher.common.domain.TokenType;
 import com.sullung2yo.seatcatcher.common.jwt.provider.JwtTokenProviderImpl;
-import com.sullung2yo.seatcatcher.user.domain.Provider;
+import com.sullung2yo.seatcatcher.domain.auth.enums.Provider;
 import com.sullung2yo.seatcatcher.user.domain.User;
 import com.sullung2yo.seatcatcher.user.domain.UserRole;
-import com.sullung2yo.seatcatcher.user.dto.request.AppleAuthRequest;
+import com.sullung2yo.seatcatcher.domain.auth.dto.request.AppleAuthRequest;
 import com.sullung2yo.seatcatcher.user.repository.UserRepository;
-import com.sullung2yo.seatcatcher.user.service.AuthServiceImpl;
+import com.sullung2yo.seatcatcher.domain.auth.service.AuthServiceImpl;
 import com.sullung2yo.seatcatcher.user.utility.random.NameGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
