+++ src/test/java/com/sullung2yo/seatcatcher/train/controller/UserTrainSeatGroupSeatControllerTest.java
@@ -3,6 +3,7 @@ package com.sullung2yo.seatcatcher.train.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sullung2yo.seatcatcher.common.domain.TokenType;
 import com.sullung2yo.seatcatcher.common.jwt.provider.JwtTokenProviderImpl;
+import com.sullung2yo.seatcatcher.domain.auth.enums.Provider;
 import com.sullung2yo.seatcatcher.train.domain.TrainSeat;
 import com.sullung2yo.seatcatcher.train.service.TrainSeatGroupService;
 import com.sullung2yo.seatcatcher.user.domain.*;
