+++ src/test/java/com/sullung2yo/seatcatcher/user/controller/TokenControllerTest.java
@@ -3,11 +3,11 @@ package com.sullung2yo.seatcatcher.user.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sullung2yo.seatcatcher.common.domain.TokenType;
 import com.sullung2yo.seatcatcher.common.jwt.provider.JwtTokenProviderImpl;
-import com.sullung2yo.seatcatcher.user.domain.Provider;
+import com.sullung2yo.seatcatcher.domain.auth.enums.Provider;
 import com.sullung2yo.seatcatcher.user.domain.User;
 
-import com.sullung2yo.seatcatcher.user.dto.request.TokenRefreshRequest;
-import com.sullung2yo.seatcatcher.user.repository.RefreshTokenRepository;
+import com.sullung2yo.seatcatcher.domain.auth.dto.request.TokenRefreshRequest;
+import com.sullung2yo.seatcatcher.domain.auth.repository.RefreshTokenRepository;
 import com.sullung2yo.seatcatcher.user.repository.UserRepository;
 import org.springframework.transaction.annotation.Transactional;
 import org.junit.jupiter.api.AfterEach;
