+++ src/test/java/com/sullung2yo/seatcatcher/user/controller/UserControllerTest.java
@@ -3,6 +3,7 @@ package com.sullung2yo.seatcatcher.user.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sullung2yo.seatcatcher.common.domain.TokenType;
 import com.sullung2yo.seatcatcher.common.jwt.provider.JwtTokenProviderImpl;
+import com.sullung2yo.seatcatcher.domain.auth.enums.Provider;
 import com.sullung2yo.seatcatcher.user.domain.*;
 import com.sullung2yo.seatcatcher.user.dto.request.UserInformationUpdateRequest;
 import com.sullung2yo.seatcatcher.user.repository.TagRepository;
