+++ src/test/java/com/sullung2yo/seatcatcher/user/repository/UserRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.sullung2yo.seatcatcher.user.repository;
 
-import com.sullung2yo.seatcatcher.user.domain.Provider;
+import com.sullung2yo.seatcatcher.domain.auth.enums.Provider;
 import com.sullung2yo.seatcatcher.user.domain.User;
 import com.sullung2yo.seatcatcher.user.domain.UserRole;
 import jakarta.persistence.EntityManager;
